# Conflicts:
#	src/Administration/Resources/app/administration/jest.config.js
#	src/Administration/Resources/app/administration/src/app/component/app/sw-app-actions/sw-app-actions.spec.js
#	src/Administration/Resources/app/administration/src/app/component/data-grid/sw-data-grid/sw-data-grid.spec.js
#	src/Administration/Resources/app/administration/src/app/component/entity/sw-category-tree-field/sw-category-tree-field.html.twig
#	src/Administration/Resources/app/administration/src/app/component/entity/sw-category-tree-field/sw-category-tree-field.scss
#	src/Administration/Resources/app/administration/src/app/component/form/select/base/sw-multi-tag-select/sw-multi-tag-select.html.twig
#	src/Administration/Resources/app/administration/src/app/component/form/sw-colorpicker-deprecated/sw-colorpicker-deprecated.spec.js
#	src/Administration/Resources/app/administration/src/app/component/form/sw-custom-field-set-renderer/sw-custom-field-set-renderer.spec.js
#	src/Administration/Resources/app/administration/test/_setup/prepare_environment.js
#	tests/acceptance/package-lock.json
#	tests/acceptance/package.json

+++ src/Core/Checkout/Cart/CartPersister.php
@@ -3,6 +3,7 @@
 namespace Shopware\Core\Checkout\Cart;
 
 use Doctrine\DBAL\Connection;
+use Psr\Clock\ClockInterface;
 use Shopware\Core\Checkout\Cart\Error\ErrorCollection;
 use Shopware\Core\Checkout\Cart\Event\CartLoadedEvent;
 use Shopware\Core\Checkout\Cart\Event\CartSavedEvent;
@@ -28,7 +29,8 @@ class CartPersister extends AbstractCartPersister
         private readonly Connection $connection,
         private readonly EventDispatcherInterface $eventDispatcher,
         private readonly CartSerializationCleaner $cartSerializationCleaner,
-        private readonly CartCompressor $compressor
+        private readonly CartCompressor $compressor,
+        private readonly ClockInterface $clock,
     ) {
     }
 
@@ -86,7 +88,11 @@ class CartPersister extends AbstractCartPersister
         $this->eventDispatcher->dispatch($event);
 
         if (!$event->shouldBePersisted()) {
-            $this->delete($cart->getToken(), $context);
+            // skipping the persistence means the stored cart stays untouched, it must not be deleted
+            if (!$cart->getBehavior()?->hasPermission(CheckoutPermissions::SKIP_CART_PERSISTENCE)) {
+                $this->delete($cart->getToken(), $context);
+                $cart->setPersisted(false);
+            }
 
             return;
         }
@@ -111,7 +117,7 @@ class CartPersister extends AbstractCartPersister
             'token' => $cart->getToken(),
             'payload' => $serializeCart,
             'rule_ids' => json_encode($context->getRuleIds(), \JSON_THROW_ON_ERROR),
-            'now' => (new \DateTime())->format(Defaults::STORAGE_DATE_TIME_FORMAT),
+            'now' => $this->clock->now()->format(Defaults::STORAGE_DATE_TIME_FORMAT),
             'compressed' => $compressed,
         ];
 
@@ -145,8 +151,7 @@ class CartPersister extends AbstractCartPersister
 
     public function prune(int $days): void
     {
-        $time = new \DateTime();
-        $time->modify(\sprintf('-%d day', $days));
+        $time = $this->clock->now()->modify(\sprintf('-%d day', $days));
 
         $stmt = $this->connection->prepare(<<<'SQL'
             DELETE FROM cart
